# Optimize encoding