--- a/modules/core/src/main/scala/mess/codec/Codec.scala
+++ b/modules/core/src/main/scala/mess/codec/Codec.scala
@@ -66,30 +66,5 @@ object Codec {
       }
   }
 
-  def serialize(m: MsgPack, acc: MessagePacker): Unit = m match {
-    case MsgPack.MEmpty     => ()
-    case MsgPack.MNil       => acc.packNil()
-    case MsgPack.MBool(a)   => acc.packBoolean(a)
-    case MsgPack.MByte(a)   => acc.packByte(a)
-    case MsgPack.MShort(a)  => acc.packShort(a)
-    case MsgPack.MInt(a)    => acc.packInt(a)
-    case MsgPack.MLong(a)   => acc.packLong(a)
-    case MsgPack.MBigInt(a) => acc.packBigInteger(a.bigInteger)
-    case MsgPack.MFloat(a)  => acc.packFloat(a)
-    case MsgPack.MDouble(a) => acc.packDouble(a)
-    case MsgPack.MString(a) => acc.packString(a)
-    case MsgPack.MExtension(t, s, a) =>
-      acc.packExtensionTypeHeader(t, s)
-      acc.writePayload(a)
-    case MsgPack.MArray(a) =>
-      acc.packArrayHeader(a.size)
-      a.foreach(serialize(_, acc))
-    case MsgPack.MMap(a) =>
-      acc.packMapHeader(a.size)
-      a.iterator.foreach {
-        case (k, v) =>
-          serialize(k, acc)
-          serialize(v, acc)
-      }
-  }
+  def serialize(m: MsgPack, acc: MessagePacker): Unit = m.pack(acc)
 }