**before**

```
[info] Benchmark                Mode  Cnt         Score        Error  Units
[info] PackBench.encodeLong10  thrpt   20    587087.901 ±   6823.606  ops/s
[info] PackBench.encodeLong30  thrpt   20    173658.439 ±   4693.170  ops/s
[info] PackBench.encodeLong60  thrpt   20     84979.790 ±    647.158  ops/s
[info] PackBench.encodeStr16   thrpt   20      1890.081 ±     23.378  ops/s
[info] PackBench.encodeStr32   thrpt   20       212.593 ±      4.016  ops/s
[info] PackBench.encodeUInt32  thrpt   20  16181031.943 ± 187892.007  ops/s
[info] PackBench.encodeUInt64  thrpt   20  15177198.338 ± 270294.091  ops/s
```


**after**

```
[info] Benchmark                Mode  Cnt         Score        Error  Units
[info] PackBench.encodeLong10  thrpt   20    663953.840 ±  26295.855  ops/s
[info] PackBench.encodeLong30  thrpt   20    197258.633 ±  12261.447  ops/s
[info] PackBench.encodeLong60  thrpt   20     98553.763 ±   2076.013  ops/s
[info] PackBench.encodeStr16   thrpt   20      1860.762 ±     32.553  ops/s
[info] PackBench.encodeStr32   thrpt   20       212.802 ±      4.494  ops/s
[info] PackBench.encodeUInt32  thrpt   20  17539041.255 ± 348184.506  ops/s
[info] PackBench.encodeUInt64  thrpt   20  17680763.769 ± 123762.091  ops/s
```